ESECLDTPLT-2661: Existing Stack Parent Template (Quickstart/Standalone)

See merge request cloudsolutions/automation-templates/f5-google-gdm-templates-v2!60

+++ automated-test-scripts/f5-google-gdm-templates-v2/common/delete_application_stack.sh
@@ -3,4 +3,9 @@
 #  replayEnabled = false
 #  replayTimeout = 0
 
-gcloud deployment-manager deployments delete <APPLICATION STACK NAME> -q
+
+if [[ "<PROVISION DEMO APP>" == "True" ]]; then
+    gcloud deployment-manager deployments delete <APPLICATION STACK NAME> -q
+else
+    echo "completed successfully"
+fi
